Issue table arrangement (#66)
* fix: adjusted position values. still need some work

* fix: adjusting table components and test data

+++ src/main/java/com/titan/table/request/TableRequest.java
@@ -15,6 +15,6 @@ public class TableRequest {
 
     private Integer tableNumber;
     private Integer numberOfPeople;
-    private Double positionX;
-    private Double positionY;
+    private Integer positionX;
+    private Integer positionY;
 }
